+++ dlatk/dlaConstants.py
@@ -55,7 +55,7 @@ DEF_MYSQL_ENGINE = 'MYISAM'
 
 ##lexInterface settings
 DEF_TERM_FIELD = 'term'
-DEF_MIN_WORD_FREQ = 1000;
+DEF_MIN_WORD_FREQ = 1000
 DEF_NUM_RAND_MESSAGES = 100
 MAX_WRITE_RECORDS = 1000 #maximum number of records to write at a time (for add_terms...)
 
@@ -78,11 +78,11 @@ DEF_MIN_FREQ = int(1) #min frequency per group to keep (don't advise above 1)
 DEF_P_OCC = float(.01) #percentage of groups a feature must appear in, to keep it
 DEF_PMI = 3.0
 DEF_MIN_FEAT_SUM = 0 #minimum sum of feature total to keep
-# DEF_STANFORD_SEGMENTER = '../Tools/StanfordSegmenter/stanford-segmenter-2014-08-27/segment.sh'
+# DEF_STANFORD_SEGMENTER = '../tools/StanfordSegmenter/stanford-segmenter-2014-08-27/segment.sh'
 DEF_STANFORD_SEGMENTER = '/home/maarten/research/tools/stanford-segmenter-2014-08-27/segment.sh'
-# DEF_STANFORD_POS_MODEL = '../Tools/StanfordTagger/stanford-postagger-2012-01-06/models/english-bidirectional-distsim.tagger' # for code release
+# DEF_STANFORD_POS_MODEL = '../tools/StanfordTagger/stanford-postagger-2012-01-06/models/english-bidirectional-distsim.tagger' # for code release
 DEF_STANFORD_POS_MODEL = '/home/hansens/Tools/StanfordTagger/stanford-postagger-2012-01-06/models/english-bidirectional-distsim.tagger'
-DEF_LEXICON_DB = 'permaLexicon'
+DEF_LEXICON_DB = 'dlatk_lexica'
 DEF_FEAT_TABLE = 'feat$1gram$messages_en$user_id$16to16$0_01'
 DEF_COLLOCTABLE = 'test_collocs'
 DEF_COLUMN_COLLOC = "feat"
@@ -136,6 +136,9 @@ DEF_OUTCOME_PATH_STARTS = []
 DEF_OUTCOME_MEDIATORS = []
 DEF_MAX_MED_SUMMARY_SIZE = 10 # maximum number of results to print in summary for each path start / outcome pair
 
+##LDA settings
+DEF_LDA_MSG_TABLE = 'messages_en_lda$msgs_en_tok_a30'
+
 ##Language filtering settings
 AVAILABLE_LANGUAGES = """
 Available:  af, am, an, ar, as, az, be, bg, bn, br,
@@ -150,7 +153,7 @@ ug, uk, ur, vi, vo, wa, xh, zh, zu"""
 DEF_SPAM_FILTER = 0.2 # threshold for removing spam users
 
 ##CoreNLP settings
-#DEF_CORENLP_DIR = '../Tools/corenlp-python' # for code release
+#DEF_CORENLP_DIR = '../tools/corenlp-python' # for code release
 DEF_CORENLP_DIR = '/home/hansens/Tools/corenlp-python'
 DEF_CORENLP_SERVER_COMMAND = './corenlp/corenlp.py'
 DEF_CORENLP_PORT = 20202   #default: 20202