+++ dlatk/lib/StanfordParser.py
@@ -22,7 +22,7 @@ import glob
 ##DEFAULTS FOR RELEASED VERSION
 # from ..dlaConstants import warn
 # try: 
-#     _InstallPath = os.path.dirname(os.path.dirname(os.path.realpath(__file__))) + '/Tools/StanfordParser/' # folder
+#     _InstallPath = os.path.dirname(os.path.dirname(os.path.realpath(__file__))) + '/tools/StanfordParser/' # folder
 #     _InstallDir = glob.glob(_InstallPath + "stanford-parser-full*")[0]
 # except:
 #     _InstallPath = ""